add, factor out the line-based parts into utility module
* Preparation for modeling the text output more nicely
