+++ lib/buchungsstreber/parser/buch_timesheet.rb
@@ -1,18 +1,18 @@
 require 'date'
 
+require_relative 'linebased_utils'
 require_relative '../entry'
 
 # BuchTimesheet parses the layout used by jo.
 #
 class Buchungsstreber::BuchTimesheet
   include Buchungsstreber::TimesheetParser::Base
+  include Buchungsstreber::TimesheetParser::LineBased
 
   def initialize(file_path, templates, minimum_time)
     @file_path = file_path
     @templates = templates
     @minimum_time = minimum_time
-
-    @model = File.readlines(@file_path) rescue []
   end
 
   def self.parses?(file)
@@ -24,7 +24,7 @@ class Buchungsstreber::BuchTimesheet
 
     current = nil
     work_hours = nil
-    @model.each do |line|
+    self.lines.each do |line|
       case line
       when /^([0-9][0-9][0-9][0-9]-[0-9][0-9]-[0-9][0-9])/
         # beginning of day
@@ -83,39 +83,6 @@ class Buchungsstreber::BuchTimesheet
     end
 
     result
-  end
-
-  def add(entries)
-    # as entries get added on top, reverse the entries before
-    entries.reverse.each do |e|
-      iso_date = e[:date].to_s
-      days = @model.map
-                   .with_index { |line, idx| [Date.parse($1), idx] if line =~ /^(\d\d\d\d-\d\d-\d\d)/ }
-                   .compact
-                   .sort { |x| x[0] }
-                   .reverse
-
-      # Find the line of the day to append to
-      idx = days.select {|x| x[0] == e[:date] }.map {|x| x[1] }.first
-
-      # or: Find the line of the day to insert new day before
-      nidx = days.select {|x| x[0] < e[:date] }.map {|x| x[1] - 1 }.first
-
-      if idx
-        @model = @model[0..idx] + [format_entry(e)] + @model[idx+1..-1]
-      elsif nidx && nidx < 0
-        @model.unshift "#{iso_date}:\n\n", format_entry(e)
-      elsif nidx
-        @model = @model[0..nidx] + ["#{iso_date}:\n", format_entry(e)] + @model[nidx+1..-1]
-      else
-        @model << "#{iso_date}\n\n"
-        @model << format_entry(e)
-      end
-    end
-  end
-
-  def unparse
-    @model.join
   end
 
   def format(entries)
